Initial Commit (2 weeks in)
While this is the first commit, this represents 2 weeks of (badly documented) work. Starting out I was using a 8x8 LED matrix driven by reading data from a MSGEQ7 and graphing the intensities of each band. That evolved into having a strip of 300 WS2812Bs, and adapting the code to have it react to the bands. However I quickly ran out of RAM on the Arduino Uno when trying to use the FastLED library, the Adafruit IO library, and drive all the LEDs which hold a lot of data and 2kB of RAM does not go far. This brought about the Teensy 4.0, which required some rewrites to enable parallel output and make sure the timings were slow enough for the strip to react.
I will be using I2S audio in the future so the MSGEQ7 code will be removed and replaced with it, and then the Adafruit IO library bits replaced with the MQTT library.