Smoothed out MQTT control, all feeds properly processed
Fully implemented the handling of new data from feeds into the global variables. Changed the simple chase to properly update when there is new data but not before then, so it is interruptable but doesnt run slow re-checking every time.

Need to re-write the other functions similarly, and finally implement the other effects fully.