Added noise effect, initial values pushed to MQTT
Made my own noise function based on the noise examples, but not using a 2d matrix. Each time the strip updates with new noise and color values it blends between the last color and the current one to provide a smoother experience. It is also interruptable and implemented with the MQTT handlers. 
Also put some publish bits in the setup to publish initial values to MQTT so that the program and the feed can have the same initial states. Since the teensy shouldnt be restarting this should happen rarely.