Teensy audio library changes
Well, this update took a lot longer to do than I wanted. I suppose for posterity's sake I should document what went wrong. 

The plan of using the teensy audio shield to get FFT and peak data was a fine theory, implementation was a whole other story. First off, I mis-soldered so many different teensy boards and shields I wanted to give up, but 3 teensys, 2 audio and Ethernet shields and plugs later, It finally worked as intended. 

Turns out that was not the last of my problems however, as the WS2812B's are both slow to update and very stringent on their timing requirements. Having the data processing from the audio shield takes up too much processing time and only allows the led strip to get the first ~40 LEDs updated before being interrupted. Now, there are various solutions like the allow_interrupts 0 flag but as it turns out, that is not supported on the Teensy 4.0. There are also other libraries like FastLED_Serial that are supposed to solve the timing issues by outputting the data as serial which, but again, I could not get it working on the 4.0. Seems the choice of board really hindered my progress here. 

My (hopeful) solution is to use the much faster, less timing specific APA102 leds, which I have ordered and should be arriving soon. Not only will I be able to push out the data way faster (to make quicker animations), the 4 wire protocol should stop the timing issues and allow the audio library to do its thing at the same time as fastled updates. 

So I did some rewrites to implement the audio library getting the "average" frequency, peak value, making a single pixel based off that, and updated the EQ and EQFromMid functions to work with the new FFT data as well as be able to be changed into an arbitrary number of channels. Without the new strip here to actually test on I am not sure if it actually works, but it compiles and works in theory.