New effects, full implementation of control variables
This was a very productive day of coding. After the APA102s came in, there were some issues in getting them hooked up without lots of flickering pixels, but I eventually managed. 

A lot of changes in this update involved re-writing the existing effects to fully use the brightness, speed, and palette variables. This took a while and a lot of uses of EVERY_N_MILLIS_I so that things can run at certain intervals, and the elimination of for loops since they cannot be controlled as easily with variable delays. EVERY_N_MILLIS_I also allows some things to run in "parallel" which can be seen in the twinkle effect where different things are happening at different intervals. 

+Twinkle effect: sets background color to a palette color, random pixels are lit up a different color before fading back
+Dissolve effect: picks a palette color, random pixels are set to that color until the whole strip matches, then picks another color and repeats
*LinearFrequency (used to be addSingleColorPixel): now takes its colors from the palette (rainbow looks best obviously), and has controllable speed which is really neat to play with
*EQ & EQFromMid: takes their colors from palettes as well, has more generic support for more channels if I want more in the future
*Bounce: Now has variable speed, uses EVERY_N_MILLIS_I so it is easily interrupted

TODO next: Add more effects, add custom palettes with more interesting/dynamic color schemes. Probably break out effect and palette definition out of main file, so the background logic/setup is one file, and the effects and palettes to be called are a separate one. Will help reduce the amount of scrolling I have to do trying to find the right section.